+++ modules/tangram-layers/src/index.js
@@ -1,5 +1,5 @@
 import {Layer} from '@deck.gl/core';
-import {Renderer} from '@vis.gl/tangram-renderer';
+import {ClassicWebGLRenderer} from '@vis.gl/tangram-renderer';
 import createTangramLayerClass, {
   getExternalCameraFrame,
   injectNextzenApiKey
@@ -9,7 +9,7 @@ import createTangramLayerClass, {
  * A deck.gl basemap layer that renders a Tangram scene into deck's active
  * luma.gl device and render pass.
  */
-const TangramLayer = createTangramLayerClass({Layer, Renderer});
+const TangramLayer = createTangramLayerClass({Layer, ClassicWebGLRenderer});
 
 export {TangramLayer, createTangramLayerClass, getExternalCameraFrame, injectNextzenApiKey};